game_analytics_pipeline/requirements.txt
# Game Analytics Pipeline - Python Dependencies
# Install: uv pip install -r requirements.txt

# ── Core ────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# ── Google Cloud ────────────────────────────────────
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.22.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
db-dtypes>=1.1.0          # BigQuery pandas dtype support
pyarrow>=12.0.0            # Parquet support for Data Locker

# ── Machine Learning ────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# ── Data Processing ─────────────────────────────────
pyarrow>=12.0.0
pandas-gbq>=0.19.0

# ── Utilities ───────────────────────────────────────
python-dotenv>=1.0.0       # .env file support
tqdm>=4.66.0               # Progress bars
tenacity>=8.2.0            # Retry logic

# ── Testing ─────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# ── Development ─────────────────────────────────────
black>=23.7.0              # Code formatting
flake8>=6.1.0              # Linting
mypy>=1.5.0                # Type checking
